Extracted string-parsing extensions into ParsePrimitivesJVM.kt
