+++ libraries/stdlib/src/kotlin/text/StringsJVM.kt
@@ -4,10 +4,9 @@
 
 package kotlin.text
 
-import java.io.StringReader
-import java.util.regex.Pattern
 import java.nio.charset.Charset
 import java.util.*
+import java.util.regex.Pattern
 
 
 /**
@@ -366,60 +365,10 @@ public inline fun String.toLowerCase(locale: java.util.Locale): String = (this a
 public inline fun String.toUpperCase(locale: java.util.Locale): String = (this as java.lang.String).toUpperCase(locale)
 
 /**
- * Returns `true` if the contents of this string is equal to the word "true", ignoring case, and `false` otherwise.
- */
-@kotlin.internal.InlineOnly
-public inline fun String.toBoolean(): Boolean = java.lang.Boolean.parseBoolean(this)
-
-/**
- * Parses the string as a signed [Byte] number and returns the result.
- * @throws NumberFormatException if the string is not a valid representation of a number.
- */
-@kotlin.internal.InlineOnly
-public inline fun String.toByte(): Byte = java.lang.Byte.parseByte(this)
-
-/**
- * Parses the string as a [Short] number and returns the result.
- * @throws NumberFormatException if the string is not a valid representation of a number.
- */
-@kotlin.internal.InlineOnly
-public inline fun String.toShort(): Short = java.lang.Short.parseShort(this)
-
-/**
- * Parses the string as an [Int] number and returns the result.
- * @throws NumberFormatException if the string is not a valid representation of a number.
- */
-@kotlin.internal.InlineOnly
-public inline fun String.toInt(): Int = java.lang.Integer.parseInt(this)
-
-/**
- * Parses the string as a [Long] number and returns the result.
- * @throws NumberFormatException if the string is not a valid representation of a number.
- */
-@kotlin.internal.InlineOnly
-public inline fun String.toLong(): Long = java.lang.Long.parseLong(this)
-
-/**
- * Parses the string as a [Float] number and returns the result.
- * @throws NumberFormatException if the string is not a valid representation of a number.
- */
-@kotlin.internal.InlineOnly
-public inline fun String.toFloat(): Float = java.lang.Float.parseFloat(this)
-
-/**
- * Parses the string as a [Double] number and returns the result.
- * @throws NumberFormatException if the string is not a valid representation of a number.
- */
-@kotlin.internal.InlineOnly
-public inline fun String.toDouble(): Double = java.lang.Double.parseDouble(this)
-
-/**
  * Encodes the contents of this string using the specified character set and returns the resulting byte array.
  */
 @kotlin.internal.InlineOnly
 public inline fun String.toByteArray(charset: Charset = Charsets.UTF_8): ByteArray = (this as java.lang.String).getBytes(charset)
-
-
 
 /**
  * Converts the string into a regular expression [Pattern] optionally
